feat: add Maryland hotels source with coordinates
- Add sources/maryland.py config for MD SDAT CAMA data
- 1,274 hotels/motels/B&Bs with room counts (81,096 total rooms)
- Includes lat/lon coordinates for PostGIS storage
- Fix batch.sql to cast lat/lon as float8 for asyncpg type inference

+++ db/queries/batch.sql
@@ -7,8 +7,8 @@
 -- Dedup on name+city. Uses DO NOTHING to handle any other unique constraint violations.
 INSERT INTO sadie_gtm.hotels (name, source, status, address, city, state, country, phone_google, category, external_id, external_id_type, location)
 VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11,
-        CASE WHEN $12 IS NOT NULL AND $13 IS NOT NULL
-             THEN ST_SetSRID(ST_MakePoint($13, $12), 4326)::geography
+        CASE WHEN $12::float8 IS NOT NULL AND $13::float8 IS NOT NULL
+             THEN ST_SetSRID(ST_MakePoint($13::float8, $12::float8), 4326)::geography
              ELSE NULL END)
 ON CONFLICT (name, city) DO UPDATE SET
     address = COALESCE(EXCLUDED.address, sadie_gtm.hotels.address),
